Internal reorg: name the core/app boundary in code
- Extract _hardening_middleware() in agent.py: the ordered hardened
  stack plus the (retry, json_validation, tool_name) trio shared into
  the context/research subagents. Stack order and instance-sharing
  topology preserved byte-for-byte; the two APP-layer members
  (WriteCollision, GitPushBlocker) are documented as shipping by
  default, removal being eval-gated.
- Move ThreadManager verbatim to assist/thread_manager.py (app-side
  persistence policy; stays in the assist package because the eval
  harness uses it without the web app). No re-export shim — all 12
  in-repo import sites updated; patch targets in TM tests updated.
  Import-time sqlite connect and the lazy model property are
  explicitly preserved (event-loop guardrail).
- Delete dead middleware: local_context_middleware.py,
  thread_logging_middleware.py (zero references).

608 unit tests pass; manage.web/cli/retention import-verified.
