+++ edd/eval/test_dev_skill_multi_turn.py
@@ -21,7 +21,7 @@ import shutil
 
 from unittest import TestCase
 
-from assist.thread import ThreadManager
+from assist.thread_manager import ThreadManager
 from assist.sandbox_manager import SandboxManager
